feat: move one more archive api
