+++ cumulus_lambda_functions/uds_api/granules_api.py
@@ -7,8 +7,6 @@ from starlette.responses import Response, JSONResponse
 
 from cumulus_lambda_functions.cumulus_wrapper.query_granules import GranulesQuery
 
-from cumulus_lambda_functions.uds_api.dapa.daac_archive_crud import DaacArchiveCrud, DaacDeleteModel, DaacAddModel, \
-    DaacUpdateModel
 from cumulus_lambda_functions.uds_api.dapa.granules_dapa_query_es import GranulesDapaQueryEs
 from cumulus_lambda_functions.lib.uds_db.granules_db_index import GranulesDbIndex
 from cumulus_lambda_functions.uds_api.fast_api_utils import FastApiUtils
@@ -204,32 +202,3 @@ async def delete_single_granule_dapa_actual(request: Request, collection_id: str
 #         response.status_code = delete_prep_result['statusCode']
 #         return delete_prep_result['body']
 #     raise HTTPException(status_code=delete_prep_result['statusCode'], detail=delete_prep_result['body'])
-
-
-@router.put("/{collection_id}/archive/{granule_id}")
-@router.put("/{collection_id}/archive/{granule_id}/")
-async def archive_single_granule_dapa(request: Request, collection_id: str, granule_id: str):
-    authorizer: UDSAuthorizorAbstract = UDSAuthorizerFactory() \
-        .get_instance(UDSAuthorizerFactory.cognito,
-                      es_url=os.getenv('ES_URL'),
-                      es_port=int(os.getenv('ES_PORT', '443'))
-                      )
-    auth_info = FastApiUtils.get_authorization_info(request)
-    collection_identifier = UdsCollections.decode_identifier(collection_id)
-    if not authorizer.is_authorized_for_collection(DBConstants.read, collection_id,
-                                                   auth_info['ldap_groups'],
-                                                   collection_identifier.tenant,
-                                                   collection_identifier.venue):
-        LOGGER.debug(f'user: {auth_info["username"]} is not authorized for {collection_id}')
-        raise HTTPException(status_code=403, detail=json.dumps({
-            'message': 'not authorized to execute this action'
-        }))
-    try:
-        api_base_prefix = FastApiUtils.get_api_base_prefix()
-        pg_link_generator = PaginationLinksGenerator(request)
-        granules_dapa_query = GranulesDapaQueryEs(collection_id, 1, None, None, filter, None, f'{pg_link_generator.base_url}/{api_base_prefix}')
-        granules_result = granules_dapa_query.archive_single_granule(granule_id)
-    except Exception as e:
-        LOGGER.exception('failed during get_granules_dapa')
-        raise HTTPException(status_code=500, detail=str(e))
-    return {'message': 'archive initiated'}
